+++ CHANGES.md
@@ -2,10 +2,13 @@
 
 ## Version 0.3.0 (in development)
 
-- Added cure rule "flags"
+- Added more rules
+  - core rule "flags"
+  - core rule "time-coordinate"  (#15)
+  - xcube rule "time-naming"  (#15)
 
 - Fixed problem where referring to values in modules via 
-  the form `"<module>:<attr>"` raised. #21
+  the form `"<module>:<attr>"` raised. (#21)
 
 - Introduced factory method `new_plugin` which simplifies
   creating plugin objects.